+++ ReleaseNotes.md
@@ -10,6 +10,7 @@ A. Start with the Quick Start guide : https://github.com/SteveGilham/altcover/wi
     * making the `ToolType` member a `Fake.DotNet.ToolType` 
     * removing the `FakeToolType` field (effectively, this and the above mean a rename/move of the `Fake` over the obsoleted `AltCover` type)
     * removing the `WithToolType` method (assign the value directly to the `ToolType` now)
+* Rename the `AltCover_Fake` namespace to `AltCoverFake`, because it's simpler that way
 * Remove the deprecated `altcover.dotnet` package with the old-style DotNetCliToolReference Tool
 
 # 6.7.xxx  (Fukurou series release 13)